Conflicts:
	bin/kaapi_event_reader/Makefile.am
	bin/kaapi_event_reader/katracereader.cpp
	src/kaapi_event_recorder.h
	src/machine/mt/kaapi_mt_init.c

+++ src/adapt/kaapi_adapt_api_lr.c
@@ -48,7 +48,7 @@
 int kaapi_api_listrequest_iterator_count(
   struct kaapi_listrequest_iterator_t* lrrange
 )
-{ return kaapi_listrequest_iterator_count(lrrange); }
+{ return (int)kaapi_listrequest_iterator_count(lrrange); }
 
 kaapi_request_t* kaapi_api_listrequest_iterator_get( 
   struct kaapi_listrequest_t* lrequests, struct kaapi_listrequest_iterator_t* lrrange
